feat: Adds content provider
Adds movies content provider with implementation of its methods.

+++ app/src/main/java/popmovies/udacity/com/model/api/MoviesApi.java
@@ -95,7 +95,7 @@ public interface MoviesApi {
      */
     @GET(MOVIE_DETAIL_API_ROUTE)
     Call<Movie> getMovieDetails(
-            @Path(MOVIE_ID_PARAMETER_NAME) Integer movieId,
+            @Path(MOVIE_ID_PARAMETER_NAME) Long movieId,
             @Query(API_KEY_QUERY_PARAMETER_NAME) String apiKey
     );
 
@@ -108,7 +108,7 @@ public interface MoviesApi {
      */
     @GET(MOVIE_REVIEWS_API_ROUTE)
     Call<ReviewsResponse> getMovieReviews(
-            @Path(MOVIE_ID_PARAMETER_NAME) Integer movieId,
+            @Path(MOVIE_ID_PARAMETER_NAME) Long movieId,
             @Query(API_KEY_QUERY_PARAMETER_NAME) String apiKey
     );
 
@@ -121,7 +121,7 @@ public interface MoviesApi {
      */
     @GET(MOVIE_VIDEOS_API_ROUTE)
     Call<VideosResponse> getMovieVideos(
-            @Path(MOVIE_ID_PARAMETER_NAME) Integer movieId,
+            @Path(MOVIE_ID_PARAMETER_NAME) Long movieId,
             @Query(API_KEY_QUERY_PARAMETER_NAME) String apiKey
     );
 }
